fix cmake dependency and move headers to include
