+++ Makefile.linux
@@ -24,7 +24,7 @@ AR = ar
 DFLAGS = -D__SYS_ZLIB
 OFLAGS =
 LFLAGS = -lbz2 -lz
-CFLAGS = -fPIC -D_7ZIP_ST
+CFLAGS = -fPIC -D_7ZIP_ST -Iinclude
 CFLAGS += $(OFLAGS) $(DFLAGS)
 
 OBJS =  src/adpcm/adpcm.o \
@@ -292,6 +292,6 @@ all: $(LIB)
 install: $(LIB)
 	install $(LIB) /usr/local/lib
 	mkdir -p /usr/local/include/StormLib
-	cp src/StormLib.h /usr/local/include/StormLib
-	cp src/StormPort.h /usr/local/include/StormLib
+	cp include/StormLib.h /usr/local/include/StormLib
+	cp include/StormPort.h /usr/local/include/StormLib
 	ldconfig
