+++ src/FileStream.cpp
@@ -13,8 +13,8 @@
 /* 11.06.10  1.00  Lad  Derived from StormPortMac.cpp and StormPortLinux.cpp */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 #include "FileStream.h"
