+++ src/SBaseCommon.cpp
@@ -11,8 +11,8 @@
 /* 06.09.10  1.01  Lad  Renamed to SBaseCommon.cpp                           */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 char StormLibCopyright[] = "StormLib v " STORMLIB_VERSION_STRING " Copyright Ladislav Zezula 1998-2014";
