+++ src/SBaseSubTypes.cpp
@@ -8,8 +8,8 @@
 /* 02.11.11  1.00  Lad  The first version of SBaseSubTypes.cpp               */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 /*****************************************************************************/
