+++ src/SFileAttributes.cpp
@@ -8,8 +8,8 @@
 /* 12.06.04  1.00  Lad  The first version of SAttrFile.cpp                   */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
