+++ src/SFileCompactArchive.cpp
@@ -10,8 +10,8 @@
 /* 21.04.13  1.02  Dea  Compact callback now part of TMPQArchive             */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 /*****************************************************************************/
