+++ src/SFileCreateArchive.cpp
@@ -9,8 +9,8 @@
 /* 08.06.10  1.00  Lad  Renamed to SFileCreateArchive.cpp                    */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
