+++ src/SFileExtractFile.cpp
@@ -8,8 +8,8 @@
 /* 20.06.03  1.00  Lad  The first version of SFileExtractFile.cpp            */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 bool WINAPI SFileExtractFile(HANDLE hMpq, const char * szToExtract, const TCHAR * szExtracted, DWORD dwSearchScope)
