+++ src/SFileFindFile.cpp
@@ -8,8 +8,8 @@
 /* 25.03.03  1.00  Lad  The first version of SFileFindFile.cpp               */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
