+++ src/SFileGetFileInfo.cpp
@@ -8,8 +8,8 @@
 /* 30.11.13  1.00  Lad  The first version of SFileGetFileInfo.cpp            */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
