+++ src/SFileOpenFileEx.cpp
@@ -8,8 +8,8 @@
 /* xx.xx.99  1.00  Lad  The first version of SFileOpenFileEx.cpp             */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 /*****************************************************************************/
