+++ src/SFilePatchArchives.cpp
@@ -8,8 +8,8 @@
 /* 18.08.10  1.00  Lad  The first version of SFilePatchArchives.cpp          */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
