+++ src/SFileReadFile.cpp
@@ -9,8 +9,8 @@
 /* 24.03.99  1.00  Lad  Added the SFileGetFileInfo function                  */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
