+++ src/SFileVerify.cpp
@@ -13,8 +13,8 @@
 /* 04.05.10  1.00  Lad  The first version of SFileVerify.cpp                 */
 /*****************************************************************************/
 
-#define __STORMLIB_SELF__
-#include "StormLib.h"
+#define __STORMLIB_NO_AUTO_LINK__
+#include <StormLib.h>
 #include "StormCommon.h"
 
 //-----------------------------------------------------------------------------
