+++ src/adpcm/adpcm.cpp
@@ -13,7 +13,7 @@
 /* 10.01.13  3.00  Lad  Refactored, beautified, documented :-)               */
 /*****************************************************************************/
 
-#include "../StormPort.h"
+#include <StormPort.h>
 #include "adpcm.h"
 
 //-----------------------------------------------------------------------------
