+++ src/huffman/huff.cpp
@@ -19,7 +19,7 @@
 #include <assert.h>
 #include <string.h>
  
-#include "../StormPort.h"
+#include <StormPort.h>
 #include "huff.h"
 
 //-----------------------------------------------------------------------------
